slim_msm: Optimize slimbus data port's connection with BAM pipe
Slimbus data port on apps side can be connected in any diretion (as
source, or as destination). BAM pipe connected to that port should
only be disconnected/reconnected if this port direction changes.
This helps with reducing calls to connect/disconnect BAM pipes.
Allocate descriptors to avoid getting descriptor-space from the
common pool and free them only when the port is deallocated.

Change-Id: I8cab16d852f345cae5e81d67d7eb0ce4d83f928d
Signed-off-by: Sagar Dharia <sdharia@codeaurora.org>