slim_msm: Add support for external MDM restart
If external MDM restarts when MDM slimbus channels are active, it
is possible that the MDM framer was clocking the bus, and ADSP was
not active. Make sure apps wakes up ADSP and checks the framer status
so that slimbus is clocked by ADSP instead of MDM, and communication
over slimbus is restored.

Change-Id: Ia75950770fa3eee3afbc8940e4b2d5437d922e9c
Signed-off-by: Naveen Kaje <nkaje@codeaurora.org>