slimbus: Provide framer-booted notification to slimbus device drivers
When slimbus resets (due to bus reset sequence, or after it shuts down
and comes up), framer boots up and starts clocking the bus.
Devices on the bus may be in undefined state at that time. Callback
is provided to slimbus device drivers so that controller can report
framer-booted notification so that drivers can take necessary action
to put devices in reset state so that they can acquire sync, report
present, and be operational.

Change-Id: I40bb8a9e3e979fa78691c690862989d48c35b408
Signed-off-by: Sagar Dharia <sdharia@codeaurora.org>